split common sources into .h/.c pairs, add paletted support
