+++ examples/common/CMakeLists.txt
@@ -30,8 +30,13 @@ endif()
 
 # a lib of common code shared between all emulators
 fips_begin_lib(common)
-    fips_files(common.c common.h)
-    fips_files(clock.h fs.h gfx.h keybuf.h prof.h)
+    fips_files(
+        common.h
+        clock.c clock.h
+        fs.c fs.h
+        gfx.c gfx.h
+        keybuf.c keybuf.h
+        prof.c prof.h)
     sokol_shader(shaders.glsl ${slang})
     if (FIPS_OSX)
         fips_files(sokol.m)
