--- examples/common/common.c
@@ -1,9 +1,0 @@
-#define COMMON_IMPL
-#include <stdint.h>
-#include <stdbool.h>
-#include "clock.h"
-#include "fs.h"
-#include "gfx.h"
-#include "keybuf.h"
-#include "prof.h"
-
